feat(review): reorganize policy classes into a dedicated package and enhance configuration clarity
