+++ processing/src/main/resources/application-processing.yaml
@@ -4,6 +4,10 @@ processing:
   similarity:
     # Minimum similarity (confidence) required for suggestions. Range: 0.0 - 1.0
     min-score: 0.6
+    # Bounds used during candidate sanitization: values outside this range are clamped.
+    # These are rarely changed; defaults correspond to valid similarity domain [0.0, 1.0].
+    min-similarity: 0.0
+    max-similarity: 1.0
 
     # When true, weight contributions by similarity (more weight to higher similarity).
     use-rank-weighting: true
@@ -25,10 +29,6 @@ processing:
     # Per-mark decay factor (0.0-1.0). When >0, additional evidences for the same mark
     # contribute with diminishing returns: multiplier = decay^(rank-1). Use 1.0 for no decay.
     per-mark-decay: 0.5
-
-    # Rank blending factor: 0.0 = use similarity only, 1.0 = use DB rank only. Small values (0.05-0.2)
-    # preserve ordering information while keeping similarity as the main signal.
-    rank-factor: 0.10
 
   # Embedding configuration
   embedding:
